Split up image values to client, server and pgsql
This change mainly comes from a need of adding a custom FetchProvider (https://github.com/permitio/opal-fetcher-postgres). Doing so requires the user to be able to build and use a different client image.

In order to keep things consistent between all three images (server, client and pgsql) this change adds all three as subfield to the `image` field with the same subfields. There are appropriate defaults for all values such that the default values.yaml produce the same spec as before the change.

This commit also removes the pullPolicy field from the schema since it was not used anywhere.